Source_code/requirements.txt
pillow
opencv-python
numpy
customtkinter
exif
piexif
auto-py-to-exe